moved utils fct of dask-sidecar to dask-sidecar
